requirements.txt
Flask==2.3.3
Jinja2==3.1.2
requests==2.31.0
PyYAML==6.0.1
markupsafe==2.1.3
itsdangerous==2.1.2
werkzeug==2.3.7